test: one sensor test module, with the shared contract parametrized
test_sensor_coverage.py was a second test module for sensor.py, opening with
'Additional tests for sensor.py to cover missing lines'. Two modules for one
source module contradicts the project's own rule, and the split was arbitrary:
the names did not overlap, so nothing marked which file a given sensor's
behaviour was described in.

Inside it, 15 classes each carried an identical test_none_when_no_data and 6
an identical test_extra_attrs_empty_when_no_data — one contract, written out
21 times. That contract ('no coordinator data, no state') is now two
parametrized tests that name it once and list the sensors it applies to, so a
new sensor joins the list instead of copying a method.

Merged into test_sensor.py; _make_optimization_coordinator now delegates to
the general _make_opt_coord rather than building its own.
